requirements.txt
fastapi
uvicorn
pillow
numpy
torch
torchvision
onnx
onnxruntime
requests
python-multipart
